project-service/requirements.txt
django==5.1.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
cryptography==43.0.3
requests==2.32.3